refactor and style changes
