+++ apps/builddao/widget/Proposals.jsx
@@ -1,0 +1,1 @@
+return <div className="text-white">Proposals</div>;
