+++ apps/builddao/widget/page/proposal.jsx
@@ -1,0 +1,8 @@
+return (
+  <div>
+    <Widget
+      src="astraplusplus.ndctools.near/widget/DAO.Proposals.index"
+      props={{ daoId: "build.sputnik-dao.near" }}
+    />
+  </div>
+);
